% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/permutation_functions.R
\name{calc_pvals}
\alias{calc_pvals}
\title{Calulate Permuted P-values}
\usage{
calc_pvals(obs_diff, perms)
}
\description{
For each density point on the grid, finds the proportion of permuted
densities which are more extreme that the observed density.
}

