% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/permutation_functions.R
\name{dens_diff}
\alias{dens_diff}
\title{Difference in Densities}
\usage{
dens_diff(npoints, data = pit_LvL)
}
\description{
Calculates the difference in pitch density when pitchers are responsible for
the men on based and when they are not responsible.
}

