% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/permutation_functions.R
\name{perm_pvals}
\alias{perm_pvals}
\title{Obtain p-values from Permutations}
\usage{
perm_pvals(B, npoints, data = pit_LvL, obs_diff = NULL)
}
\description{
Wrapper which uses \code{\link{dens_diff}}, \code{\link{perm_diffs}}, and
\code{\link{calc_pvals}} to obtain p-values at each grid point.
}

