% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/permutation_functions.R
\name{pvals_to_long}
\alias{pvals_to_long}
\title{Turns matrix of p-values into a \code{\link{tbl}} into long format.}
\usage{
pvals_to_long(dat)
}
\arguments{
\item{dat}{matrix of p-values.  Each row/column combination corresponds to a
horizontal/vertical location in the strikezone.  Obtained from
\code{\link{perm_pvals}}.}
}
\description{
A nice way to turn a matrix of p-values into long format.
}

