% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/permutation_functions.R
\name{perm_diffs}
\alias{perm_diffs}
\title{Permuted Differences}
\usage{
perm_diffs(B, npoints, data = pit_LvL)
}
\description{
Permutes the data B times and calculates the difference in densities for each
of the B permutations.
}

