% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/permutation_functions.R
\name{perm_pvals}
\alias{perm_pvals}
\title{Obtain p-values from Permutations}
\usage{
perm_pvals(B, npoints, data = pit_LvL, obs_diff = NULL, perm_dens = NULL)
}
\arguments{
\item{B}{the number of permutations to run}

\item{npoints}{the number of grid points in each direction.  Densities are
calculated at each one of these grid points.}

\item{data}{the data.frame containing pitch location and pitcher
responsibility.  Defaults to \code{pit_LvL}.}

\item{obs_diff}{The observed difference in densities.  If specified manually,
should come from \code{\link{obs_diff}}.  Defaults to NULL, in which case
it will be calculated using \code{\link{obs_diff}}.}

\item{perm_dens}{The permuted densities.  If specified, must be a three
dimensional array returned from \code{\link{perm_dens}}.  Defaults to NULL,
in which case it will be calculated using \code{\link{perm_dens}}.}
}
\description{
Wrapper which uses \code{\link{dens_diff}}, \code{\link{perm_diffs}}, and
\code{\link{calc_pvals}} to obtain p-values at each grid point.
}

